requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo[srv]==3.6
pymongo

-e .